+++ test_switch.py
@@ -17,7 +17,7 @@ class MockPlayer:
         return others[0]
 
 
-def __mock_setup_round__(hands, stock, discards,
+def mock_setup_round(hands, stock, discards,
                          skip=False, draw2=False, draw4=False, direction=1):
     def str_to_cards(spec):
         return [Card(sv[:1], sv[1:]) for sv in spec.split()]
@@ -33,7 +33,7 @@ def __mock_setup_round__(hands, stock, discards,
     return s
 
 
-def __test_setup_round__resets_flags__():
+def test_setup_round__resets_flags():
     """setup_round sets all flags to initial value"""
     s = switch.Switch()
     s.players = [MockPlayer([]), MockPlayer([])]
@@ -44,7 +44,7 @@ def __test_setup_round__resets_flags__():
     assert s.direction == 1
 
 
-def __test_setup_round__deals_cards__():
+def test_setup_round__deals_cards():
     """setup_round deals correct number of cards"""
     s = switch.Switch()
     s.players = [MockPlayer([]), MockPlayer([])]
@@ -54,9 +54,9 @@ def __test_setup_round__deals_cards__():
     assert len(s.stock) == 52 - len(s.players) * 7 - 1
 
 
-def __test_pick_up_card__pick_correct_number__():
+def test_pick_up_card__pick_correct_number():
     """pick_up_card picks up correct number of cards"""
-    s = __mock_setup_round__(['♣4', '♣9'], '♠7 ♢8 ♠5 ♢6 ♢7', '♠5 ♢6 ♡3')
+    s = mock_setup_round(['♣4', '♣9'], '♠7 ♢8 ♠5 ♢6 ♢7', '♠5 ♢6 ♡3')
     player = s.players[0]
     picked = s.__pick_up_card__(player, 4)
     assert picked == 4
@@ -69,8 +69,8 @@ def __test_pick_up_card__pick_correct_number__():
     assert len(s.stock) == 0
 
 
-def __test_can_discard__follows_suit__():
-    s = __mock_setup_round__([], '', '♣5')
+def test_can_discard__follows_suit():
+    s = mock_setup_round([], '', '♣5')
     assert s.__can_discard__(Card('♣', '2'))
     assert s.__can_discard__(Card('♣', '3'))
     assert s.__can_discard__(Card('♣', '4'))
@@ -83,85 +83,85 @@ def __test_can_discard__follows_suit__():
     assert s.__can_discard__(Card('♣', 'K'))
 
 
-def __test_can_discard__follows_value__():
-    s = __mock_setup_round__([], '', '♣5')
+def test_can_discard__follows_value():
+    s = mock_setup_round([], '', '♣5')
     assert s.__can_discard__(Card('♢', '5'))
     assert s.__can_discard__(Card('♡', '5'))
     assert s.__can_discard__(Card('♠', '5'))
 
 
-def __test_can_discard__allows_ace__():
-    s = __mock_setup_round__([], '', '♣5')
+def test_can_discard__allows_ace():
+    s = mock_setup_round([], '', '♣5')
     assert s.__can_discard__(Card('♢', 'A'))
     assert s.__can_discard__(Card('♡', 'A'))
     assert s.__can_discard__(Card('♠', 'K'))
 
 
-def __test_can_discard__allows_queen__():
-    s = __mock_setup_round__([], '', '♣5')
+def test_can_discard__allows_queen():
+    s = mock_setup_round([], '', '♣5')
     assert s.__can_discard__(Card('♢', 'Q'))
     assert s.__can_discard__(Card('♡', 'Q'))
     assert s.__can_discard__(Card('♠', 'Q'))
 
 
-def __test_discard_card__sets_skip__():
-    s = __mock_setup_round__(['♣4 ♡8', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
+def test_discard_card__sets_skip():
+    s = mock_setup_round(['♣4 ♡8', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
     s.__discard_card__(s.players[0], Card('♡', '8'))
     assert s.skip
 
 
-def __test_discard_card__sets_draw2__():
-    s = __mock_setup_round__(['♣4 ♡2', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
+def test_discard_card__sets_draw2():
+    s = mock_setup_round(['♣4 ♡2', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
     s.__discard_card__(s.players[0], Card('♡', '2'))
     assert s.draw2
 
 
-def __test_discard_card__sets_draw4__():
-    s = __mock_setup_round__(['♣4 ♡Q', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
+def test_discard_card__sets_draw4():
+    s = mock_setup_round(['♣4 ♡Q', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
     s.__discard_card__(s.players[0], Card('♡', 'Q'))
     assert s.draw4
 
 
-def __test_discard_card__reverses__():
-    s = __mock_setup_round__(['♣4 ♡K', '♣K ♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
+def test_discard_card__reverses():
+    s = mock_setup_round(['♣4 ♡K', '♣K ♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
     s.__discard_card__(s.players[0], Card('♡', 'K'))
     assert s.direction == -1
     s.__discard_card__(s.players[1], Card('♣', 'K'))
     assert s.direction == 1
 
 
-def __test_discard_card__swaps__():
-    s = __mock_setup_round__(['♣4 ♡J', '♣K ♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
+def test_discard_card__swaps():
+    s = mock_setup_round(['♣4 ♡J', '♣K ♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
     s.__discard_card__(s.players[0], Card('♡', 'J'))
     assert s.players[0].hand == [Card('♣', 'K'), Card('♣', '9')]
     assert s.players[1].hand == [Card('♣', '4')]
 
 
-def __test_get_normalized_hand_sizes__():
+def test_get_normalized_hand_sizes():
     """test hand size normalization"""
-    s = __mock_setup_round__(['♣4', '♣K ♣9', '♡J ♢5 ♢6'], '♢7 ♢8', '♡3')
+    s = mock_setup_round(['♣4', '♣K ♣9', '♡J ♢5 ♢6'], '♢7 ♢8', '♡3')
     assert s.__get_normalized_hand_sizes__(s.players[0]) == [1, 2, 3]
     assert s.__get_normalized_hand_sizes__(s.players[1]) == [2, 3, 1]
     assert s.__get_normalized_hand_sizes__(s.players[2]) == [3, 1, 2]
 
-    s = __mock_setup_round__(['♣4', '♣K ♣9', '♡J ♢5 ♢6'], '♢7 ♢8', '♡3', direction=-1)
+    s = mock_setup_round(['♣4', '♣K ♣9', '♡J ♢5 ♢6'], '♢7 ♢8', '♡3', direction=-1)
     assert s.__get_normalized_hand_sizes__(s.players[0]) == [1, 3, 2]
     assert s.__get_normalized_hand_sizes__(s.players[1]) == [2, 1, 3]
     assert s.__get_normalized_hand_sizes__(s.players[2]) == [3, 2, 1]
 
 
-def __test_swap_hands__():
+def test_swap_hands():
     """Test swapping of hands"""
-    s = __mock_setup_round__(['♣4', '♣K ♣9', '♡J ♢5 ♢6'], '♢7 ♢8', '♡3')
+    s = mock_setup_round(['♣4', '♣K ♣9', '♡J ♢5 ♢6'], '♢7 ♢8', '♡3')
     s.__swap_hands__(s.players[1], s.players[2])
     assert len(s.players[1].hand) == 3
     assert len(s.players[2].hand) == 2
 
 
-def __test_run_player__adheres_to_skip_flag__():
+def test_run_player__adheres_to_skip_flag():
     """run_player adheres to switch.skip"""
     from copy import deepcopy
-    s = __mock_setup_round__(['', '', ''], '', '♣3', skip=True)
+    s = mock_setup_round(['', '', ''], '', '♣3', skip=True)
     player = s.players[1]
     hand_before = deepcopy(player.hand)
     s.__run_player__(player)
@@ -169,35 +169,35 @@ def __test_run_player__adheres_to_skip_flag__():
     assert not s.skip
 
 
-def __test_run_player__adheres_to_draw2_flag__():
+def test_run_player__adheres_to_draw2_flag():
     """run_player adheres to switch.draw2"""
-    s = __mock_setup_round__(['', ''], '♢5 ♣6 ♣7', '♢3', draw2=True)
+    s = mock_setup_round(['', ''], '♢5 ♣6 ♣7', '♢3', draw2=True)
     player = s.players[1]
     s.__run_player__(player)
     assert len(player.hand) == 2
     assert not s.draw2
 
 
-def __test_run_player__adheres_to_draw4_flag__():
+def test_run_player__adheres_to_draw4_flag():
     """run_player adheres to switch.draw4"""
-    s = __mock_setup_round__(['', ''], '♢5 ♣5 ♣6 ♣7 ♣8', '♢3', draw4=True)
+    s = mock_setup_round(['', ''], '♢5 ♣5 ♣6 ♣7 ♣8', '♢3', draw4=True)
     player = s.players[1]
     s.__run_player__(player)
     assert len(player.hand) == 4
     assert not s.draw4
 
 
-def __test_run_player__returns_true_upon_win__():
+def test_run_player__returns_true_upon_win():
     """run_player returns True if player wins"""
-    s = __mock_setup_round__(['♣4 ♣5', '♣9', '♣10'], '♣6 ♣7 ♣8', '♣3')
+    s = mock_setup_round(['♣4 ♣5', '♣9', '♣10'], '♣6 ♣7 ♣8', '♣3')
     player = s.players[0]
     assert not s.__run_player__(player)
     assert s.__run_player__(player)
 
 
-def __test_run_player__draws_card__():
+def test_run_player__draws_card():
     """run_player forces pick up if no discard possible"""
-    s = __mock_setup_round__(['♣4', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
+    s = mock_setup_round(['♣4', '♣9'], '♢5 ♢6 ♢7 ♢8', '♡3')
     player = s.players[0]
     s.__run_player__(player)
     assert len(player.hand) == 2
@@ -205,9 +205,9 @@ def __test_run_player__draws_card__():
     assert len(s.discards) == 1
 
 
-def __test_run_player__draws_card_and_discards__():
+def test_run_player__draws_card_and_discards():
     """run_player discards drawn card if possible"""
-    s = __mock_setup_round__(['♣4', '♣9'], '♢5 ♢6 ♢7 ♡8', '♡3')
+    s = mock_setup_round(['♣4', '♣9'], '♢5 ♢6 ♢7 ♡8', '♡3')
     player = s.players[0]
     s.__run_player__(player)
     assert len(player.hand) == 1