+++ CHANGELOG.md
@@ -31,6 +31,9 @@ name should be lowercase.
 * v1.1.9 [2019-12-03]: Fixed name python function name shadowing in file user_interface.py. Used "minimum" instead of
 "min" and maximum instead of "max".
 
+* v1.1.10 [2019-12-03]: Fixed player index out of bounds error to "i = self.direction % len(self.players) - i" in
+switch.py. test_switch updated some methods privacy
+